new file:   autumn/number-to-bytes/go/d.go

+++ autumn/number-to-bytes/go/b.go
@@ -9,8 +9,8 @@ import (
 func main() {
    var (
       b = new(bytes.Buffer)
-      n uint8 = 0xFF
+      n uint16 = 0xFF
    )
    binary.Write(b, binary.BigEndian, n)
-   fmt.Println(b.Bytes()) // [255]
+   fmt.Println(b.Bytes()) // [0 255]
 }
